New subcommand technique (redesigned exec in cli.go)
This commit includes a rewrite for 'runCommand()', the function
which is used to run local subcommands. The rewrite fixes
some long-standing bugs related to streaming the output of a subcommand
to the console as the command is running (like perl's 'system()' behaviour).

The previous implementation (using Scanner) would freeze if the subcommand
used the carriage return ('\r') character, which npm (among others) uses to simulate a
loading bar. This implementation doesn't freeze in that case, but it
considers each line of output to be a new line (ignoring npm's desire to
make it look like one line of dynamic output).

This commit also includes some changes as introduced by gofmt, since my editor
won't let me turn that off anymore.