Fixes a bug in the dirty workdir check
The old line looked like this:
    getCommandExitCode("test", "-z \"$(git ls-files --others)\"")

Here's the problem - bash subcommands can't be used in this format of
command exec, where the args are passed directly to the exec binary (git,
in this case). This let the command to fail each time.

So, the new line looks like this:
    getCommandOutput("git", "ls-files --others") == ""

Which is a little less elegant (checking against the empty string), but
it's effectively what the 'test' command was doing anyway.